Update rules link again for new rule docs
This shouldn't be merged yet, but is the update we'll need to this link again when this change gets released: https://github.com/sqlfluff/sqlfluff/pull/4695

The old link will still "work" (i.e. it won't be a broken link), but it will stop pointing to the right point on the page. I've added some more permanent redirect stubs so we can support more consistent links even if we change the format again in the future.